Reorganize code storaging
